Made data class from class and replaced int with long in ID

+++ Android. Module 10. Async operations/app/src/main/java/com/example/asyncoperations/model/room/entities/Character.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Character(
     @PrimaryKey
-    val idCharacter: Int,
+    val idCharacter: Long,
     val image: String?,
     val name: String?,
     val species: String?,
